@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, Component, Input, ViewEncapsulation } from '@angular/core';
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 import { ProductsWithRatingQueryModel } from 'src/app/query-models/products-with-rating.query-model';
 import { SortingValueQueryModel } from 'src/app/query-models/sorting-value.query-model';
 
@@ -10,7 +10,7 @@ import { SortingValueQueryModel } from 'src/app/query-models/sorting-value.query
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class SelectedSortingValueComponent {
-  @Input() selectedSortingValue!: FormControl
+  @Input() selectedSortingValue!: UntypedFormControl
   @Input() filteredProducts!: ProductsWithRatingQueryModel[]
   @Input() sortingValues!: SortingValueQueryModel[]
 
