@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, Component, ViewEncapsulation } from '@angular/core';
-import { FormControl, FormGroup } from '@angular/forms';
+import { UntypedFormControl, UntypedFormGroup } from '@angular/forms';
 import { ActivatedRoute, Params } from '@angular/router';
 import { Observable, combineLatest } from 'rxjs';
 import { filter, map, startWith, switchMap } from 'rxjs/operators';
@@ -26,8 +26,8 @@ export class StoreProductsComponent {
   readonly oneStore$: Observable<StoreModel> = this.activatedRouteParam$.pipe(
     switchMap(data => this._storeService.getOne(data['storeId'])))
 
-  readonly form: FormGroup = new FormGroup({
-    search: new FormControl('')
+  readonly form: UntypedFormGroup = new UntypedFormGroup({
+    search: new UntypedFormControl('')
   });
 
   readonly productsInStore$: Observable<ProductModel[]> = combineLatest([
